refactor: move completeness into packages
