+++ README.md
@@ -21,10 +21,10 @@ We'll provide examples for iron.io functions, google cloud functions and AWS Lam
 
 ## Packages
 
-| Package | Version | Dependencies |
-|--------|-------|------------|
+| Package                                            | Version                                                                           | Dependencies                                                                                                    |
+| -------------------------------------------------- | --------------------------------------------------------------------------------- | --------------------------------------------------------------------------------------------------------------- |
 | [`example-inventory`](/packages/example-inventory) | [![example-inventory Version][example-inventory-icon]][example-inventory-version] | [![example-inventory Dependencies Status][example-inventory-dependencies-icon]][example-inventory-dependencies] |
-| [`extension-base`](/packages/extension-base) | [![extension-base Version][extension-base-icon]][extension-base-version] | [![extension-base Dependencies Status][extension-base-dependencies-icon]][extension-base-dependencies] |
+| [`extension-base`](/packages/extension-base)       | [![extension-base Version][extension-base-icon]][extension-base-version]          | [![extension-base Dependencies Status][extension-base-dependencies-icon]][extension-base-dependencies]          |
 
 [example-inventory-version]: https://www.npmjs.com/package/@commercetools/example-inventory
 [example-inventory-icon]: https://img.shields.io/npm/v/@commercetools/example-inventory.svg?style=flat-square
