+++ packages/example-completeness/.gcloudignore
