+++ packages/example-completeness/.gitignore
