+++ packages/example-completeness/config/completeProduct.json
