+++ packages/example-completeness/test/resources/completeProduct.json
