+++ packages/example-completeness/test/resources/incompleteProduct.json
