+++ packages/example-completeness/test/resources/normalizedCompleteProduct.json
