+++ packages/example-completeness/test/resources/product.json
